requirements.txt
# Core
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LLM
openai>=1.0.0

# Vector DB
pinecone-client>=3.0.0

# Tracing (optional)
langchain>=0.1.0
langsmith>=0.1.0

# Utilities
python-dotenv>=1.0.0
